Integrate bootstrap to home
Change the way that a reserva is made.

Started the process to add elastic search for the sessions

Started the process to chose a especific chair on the theater,
just by clicking. That means the user will no longer write the
chair number, and will be able to see which one is available or not.